+++ cmd/storagenode-updater/restart.go
@@ -1,45 +1,30 @@
-// Copyright (C) 2019 Storj Labs, Inc.
+// Copyright (C) 2020 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-// +build !unittest
+// +build unittest !windows
 
 package main
 
 import (
-	"fmt"
+	"context"
 	"os"
-	"os/exec"
-	"path/filepath"
-	"runtime"
 
+	"github.com/spf13/cobra"
 	"github.com/zeebo/errs"
 )
 
-func restartService(name string) error {
-	switch runtime.GOOS {
-	case "windows":
-		// TODO: cleanup temp .bat file
-		restartSvcBatPath := filepath.Join(os.TempDir(), "restartservice.bat")
-		restartSvcBat, err := os.Create(restartSvcBatPath)
-		if err != nil {
-			return err
-		}
+func cmdRestart(cmd *cobra.Command, args []string) error {
+	return nil
+}
 
-		restartStr := fmt.Sprintf("net stop %s && net start %s", name, name)
-		_, err = restartSvcBat.WriteString(restartStr)
-		if err != nil {
-			return err
-		}
-		if err := restartSvcBat.Close(); err != nil {
-			return err
-		}
-
-		out, err := exec.Command(restartSvcBat.Name()).CombinedOutput()
-		if err != nil {
-			return errs.New("%s", string(out))
-		}
-	default:
-		return nil
+func restartService(ctx context.Context, service, binaryLocation, newVersionPath, backupPath string) error {
+	if err := os.Rename(binaryLocation, backupPath); err != nil {
+		return errs.Wrap(err)
 	}
+
+	if err := os.Rename(newVersionPath, binaryLocation); err != nil {
+		return errs.Combine(err, os.Rename(backupPath, binaryLocation), os.Remove(newVersionPath))
+	}
+
 	return nil
 }
