--- cmd/storagenode-updater/restart_mock.go
@@ -1,8 +1,0 @@
-// Copyright (C) 2019 Storj Labs, Inc.
-// See LICENSE for copying information.
-
-// +build unittest
-
-package main
-
-func restartService(name string) error { return nil }
